[ch-#155507451] Create PostgresSQL/SQLAlchemy database with table models
