+++ .travis.yml
@@ -7,5 +7,9 @@ python:
 install:
   - pip install
   - pip install -r requirements.txt
+  - pip install Flask-Alchemy
+
+before_script:
+  - psql -c 'create database test;' -U postgres
 # command to run tests
 script: cd ./api && nosetests
